requirements.txt
fastapi==0.115.12
joblib==1.5.1
numpy==2.3.0
pydantic==2.11.7
pydantic_core==2.33.2
pytest==8.4.1
requests==2.32.4
scikit-learn==1.7.0
uvicorn==0.34.3
python-dotenv==1.1.1
httpx==0.28.1